jarvis/requirements.txt
faster-whisper==1.1.1
numpy==1.26.4
openai==1.76.0
openwakeword==0.6.0
PyOpenAL==0.7.11a1
pyaudio==0.2.14
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sounddevice==0.5.1
soundfile==0.12.1
